[iOS-custom] separate pipeline storage for faster searching rather than searching hashtable
